Improve EL affinity in multi event loop connection pool
Currently, when using several event loops for a connection pool, the provider picks event loops one after the order.

So, when a new connection is created, it's almost never the event loop that tries to acquire a connection that is used.

With this change, event loop affinity is preferred, while we continue to spread new connections evenly across available event loops.

Signed-off-by: Thomas Segismont <tsegismont@gmail.com>